Separate files with tokenizer and feature extractor
