+++ src/main/kotlin/org/jetbrains/research/commentupdater/CodeCommentTokenizer.kt
@@ -114,93 +114,3 @@ object CodeCommentTokenizer {
 
 
 
-
-object MethodFeaturesExtractor{
-    fun extractFullyQualifiedName(method: PsiMethod): String {
-        val classes = mutableListOf<String>()
-        classes.add(method.name)
-        var upperClass = method.containingClass
-        while (upperClass != null) {
-            classes.add(upperClass.name ?: "")
-            upperClass = upperClass.containingClass
-        }
-        return classes.reversed().joinToString(separator = ".")
-    }
-
-    /**
-     * Remove Comments from Method text
-     */
-    fun extractCode(method: PsiMethod): String {
-        return method.children.filter { it !is PsiComment }.joinToString (separator = "", transform={ it.text })
-    }
-
-    fun extractArguments(method: PsiMethod): List<Pair<String, String>> {
-        return method.parameterList.parameters.map {
-            it.type.presentableText to it.name
-        }
-    }
-
-    fun extractReturnType(method: PsiMethod): String {
-        return method.returnType?.presentableText ?: ""
-    }
-
-    fun extractReturnStatements(method: PsiMethod): List<String> {
-        return PsiTreeUtil.findChildrenOfAnyType(method, PsiReturnStatement::class.java)
-            .map {
-               it.returnValue?.text ?: ""
-            }.filter { it != "" }
-    }
-
-    fun extractMethodFeatures(method: PsiMethod, subTokenize: Boolean): HashMap<String, List<String>> {
-        val arguments = extractArguments(method)
-        val returnType = extractReturnType(method)
-        val returnStatements = extractReturnStatements(method)
-
-        fun String.tokens(): List<String> {
-            return if (subTokenize) {
-                CodeCommentTokenizer.subTokenizeCode(this)
-            } else {
-                CodeCommentTokenizer.tokenizeCode(this)
-            }
-        }
-        
-        return hashMapOf(
-            "argument_name" to arguments.map {
-                it.second.tokens()
-            }.flatten(),
-            "argument_type" to arguments.map {
-                it.first.tokens()
-            }.flatten(),
-            "return_type" to returnType.tokens(),
-            "return_statements" to returnStatements.map {
-                it.tokens()
-            }.flatten(),
-            "method_name" to method.name.tokens()
-        )
-
-    }
-
-}
-
-fun main() {
-    println(CodeCommentTokenizer.tokenizeText("/**\n" +
-            "   * Finds the first JavaDoc tag with the specified name.\n" +
-            "   * @param name The name of the tags to find (not including the leading @ character).\n" +
-            "   * @return the tag with the specified name, <some_html_tag> <some other tag> or null if not found.\n" +
-            "   */"))
-    println(CodeCommentTokenizer.tokenizeText("I am here and MainCaptain is here alsoChat"))
-    println(CodeCommentTokenizer.subTokenizeText("I am here and MainCaptain is here alsoChat"))
-    println(CodeCommentTokenizer.subTokenizeCode("""public class VowelConsonant {
-
-    public static void main(String[] args) {
-        // comment
-        char ch = 'i';
-        if(ch == 'a' || ch == 'e' || ch == 'i' || ch == 'o' || ch == 'u' )
-            System.out.println(ch + " is vowel");
-        else
-            System.out.println(ch + " is consonant");
-
-    }
-}"""))
-
-}
